requirements.txt
django==4.*
djangorestframework==3.*
django-filter==2.*
pillow==9.*
